LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/abort-after-send-expected.txt


FAIL XMLHttpRequest: abort() after send() assert_equals: expected "abort(0,0,false)" but got "upload.progress(0,0,false)"
